requirements.txt
python-telegram-bot
httpx
selenium